requirements.txt
flask
google
pandas
xlrd
xlsxwriter
textblob
beautifulsoup4
requests
pdfminer.six